#94 output window restrict mouse scroll testcases
