+++ goddo_player/preview_window/preview_window_output.py
@@ -363,7 +363,7 @@ class FrameInOutSlider(ClickSlider):
         painter.end()
 
     def wheelEvent(self, event: QWheelEvent) -> None:
-        logging.info('wheel event')
+        logging.debug('wheel event')
         # super().wheelEvent(e)
         if event.angleDelta().y() > 0:
             frame_diff = self.get_wheel_skip_time() * -1
