+++ goddo_test/test_preview_output_playback.py
@@ -27,390 +27,6 @@ def test_slider_disabled_initially(windows_container: WindowsContainer):
     assert windows_container.output_window.preview_widget.frame_pixmap is None
     assert windows_container.output_window.slider.value() == 0
 
-def test_unrestricted_move_with_mouse_scroll(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * 130
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:51.00')
-    go_to_prev_wind_slider(windows_container.output_window, 0.1)
-
-    # try to scroll pass start frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame)
-
-    # when exactly on start frame, try to scroll back
-    pyautogui.scroll(1)
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame)
-
-    # scroll forward when on start frame
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame + 4 * 5)
-
-    # scroll back and forth when between start and in frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:55.00')
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 55)
-
-    # scroll back and forth on in frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:55.00')
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame + 4*5)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame - 4*5)
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:02.00')
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame - 4*3)
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame + 4*2)
-
-    # # scroll backward when between in and out frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:30.00')
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 95)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 90)
-
-    # scroll back and forth on out frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:55.00')
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame + 4*5)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame - 4*5)
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:02.00')
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame - 4*3)
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame + 4*2)
-
-    # scroll back and forth when between out and end frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:08.00')
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame + 4*3)
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame + 4*8)
-
-    # stops at end frame
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame)
-
-    # when exactly on out frame, try to scroll forward
-    pyautogui.scroll(-1)
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame)
-
-    # scroll backward when on out frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame - 4*5)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    expected_cur_frame = expected_end_frame - 4*5
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.93, 0.94), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_move_with_mouse_click(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    before_frame_no = app_thread.mon.state.preview_window_output.current_frame_no
-    before_slider_value = app_thread.mon.preview_window_output.slider.value()
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * 130
-
-    assert expected_in_frame <= before_frame_no <= expected_out_frame
-
-    click_on_prev_wind_slider(windows_container.output_window, 0.1, should_slider_value_change=False)
-    time.sleep(0.5)
-    assert app_thread.mon.state.preview_window_output.current_frame_no == before_frame_no
-    assert app_thread.mon.preview_window_output.slider.value() == before_slider_value
-
-    click_on_prev_wind_slider(windows_container.output_window, 0.9, should_slider_value_change=False)
-    time.sleep(0.5)
-    assert app_thread.mon.state.preview_window_output.current_frame_no == before_frame_no
-    assert app_thread.mon.preview_window_output.slider.value() == before_slider_value
-
-    click_on_prev_wind_slider(windows_container.output_window, 0.5, should_slider_value_change=False)
-    wait_until(lambda: expected_in_frame <= app_thread.mon.state.preview_window_output.current_frame_no <= expected_out_frame)
-    assert 99 <= app_thread.mon.preview_window_output.slider.value() <= 101
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.45, 0.55), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-# unrestricted mouse click
-#   click between start and in frame, should go there
-#   click between in/out frame, should go there
-#   click between out and end frame, should go there
-#   try to drag mouse pass in frame
-#   try to drag mouse pass start frame
-#   try to drag mouse pass out frame
-#   try to drag mouse pass end frame
-def test_unrestricted_mouse_click_before_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    before_frame_no = app_thread.mon.state.preview_window_output.current_frame_no
-    before_slider_value = app_thread.mon.preview_window_output.slider.value()
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * 130
-
-    #   click between start and in frame
-    click_on_prev_wind_slider(windows_container.output_window, 0.01)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0, 0.03), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_mouse_click_between_in_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    before_frame_no = app_thread.mon.state.preview_window_output.current_frame_no
-    before_slider_value = app_thread.mon.preview_window_output.slider.value()
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * 130
-
-    #   click between start and in frame
-    click_on_prev_wind_slider(windows_container.output_window, 0.5)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.49, 0.51), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_mouse_click_after_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    before_frame_no = app_thread.mon.state.preview_window_output.current_frame_no
-    before_slider_value = app_thread.mon.preview_window_output.slider.value()
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * 130
-
-    #   click between start and in frame
-    click_on_prev_wind_slider(windows_container.output_window, 0.99)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.97, 1), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_mouse_drag_left_within_in_out(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.7)
-    x2_offset = int(slider.width() * 0.3)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.28, 0.32), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_mouse_drag_right_within_in_out(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.3)
-    x2_offset = int(slider.width() * 0.7)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.68, 0.72), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_mouse_drag_left_pass_in(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.5)
-    x2_offset = int(slider.width() * 0.05)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.12, 0.15), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_mouse_drag_right_pass_out(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.5)
-    x2_offset = int(slider.width() * 0.95)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.85, 0.88), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_mouse_drag_right_within_in_out(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.3)
-    x2_offset = int(slider.width() * 0.7)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.68, 0.72), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_mouse_drag_left_pass_in(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    slider = app_thread.mon.preview_window_output.slider
-    pos = local_to_global_pos(slider, app_thread.mon.preview_window_output)
-    x1_offset = int(slider.width() * 0.5)
-    x2_offset = int(slider.width() * 0.05)
-    y_offset = int(slider.height() * 0.5)
-    drag_and_drop(pos.x() + x1_offset, pos.y() + y_offset, pos.x() + x2_offset, pos.y() + y_offset)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    assert app_thread.mon.state.preview_window_output.current_frame_no <= 4 * 60 * 2
-    assert app_thread.mon.preview_window_output.slider.value() <= 176
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.12, 0.15), is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
 def test_video_stops_at_out_with_restrict_with_playing(app_thread, windows_container: WindowsContainer, blank_state):
     video_path = get_blank_1hr_vid_path()
     drop_video_on_preview(app_thread, windows_container, video_path)
